+++ pushapkscript/src/mozapkpublisher/common/apk/checker.py
@@ -6,8 +6,7 @@ from mozilla_version.gecko import FennecVersion
 
 from mozapkpublisher.common.apk.history import get_expected_combos, craft_combos_pretty_names
 from mozapkpublisher.common.exceptions import BadApk, BadSetOfApks, NotMultiLocaleApk
-from mozapkpublisher.common.googleplay import is_package_name_nightly
-from mozapkpublisher.common.utils import filter_out_identical_values, PRODUCT
+from mozapkpublisher.common.utils import filter_out_identical_values
 
 logger = logging.getLogger(__name__)
 
@@ -16,65 +15,55 @@ logger = logging.getLogger(__name__)
 _ARCHITECTURE_ORDER_REGARDING_VERSION_CODE = ('armeabi-v7a', 'arm64-v8a', 'x86')
 
 
-def cross_check_apks(apks_metadata_per_paths):
+class ExpectedPackageNamesCheck:
+    def __init__(self, expected_product_types):
+        self.expected_product_types = expected_product_types
+
+    def validate(self, apks_metadata_per_paths):
+        types = set([metadata['package_name'] for metadata in apks_metadata_per_paths.values()])
+
+        if not types == set(self.expected_product_types):
+            raise BadSetOfApks('Expected product types {}, found {}'.format(self.expected_product_types, types))
+        logger.info('Expected product types {} found'.format(self.expected_product_types))
+
+
+class AnyPackageNamesCheck:
+    def validate(self, _):
+        pass
+
+
+def cross_check_apks(apks_metadata_per_paths, package_names_check, skip_checks_fennec, skip_check_multiple_locales,
+                     skip_check_same_locales, skip_check_ordered_version_codes):
     logger.info("Checking APKs' metadata and content...")
-    package_name = list(apks_metadata_per_paths.values())[0]['package_name']
-    if PRODUCT.is_focus_flavor(package_name):
-        cross_check_focus_apks(apks_metadata_per_paths)
-    elif PRODUCT.is_reference_browser(package_name):
-        cross_check_reference_browser_apks(apks_metadata_per_paths)
-    elif PRODUCT.is_fenix(package_name):
-        cross_check_fenix_apks(apks_metadata_per_paths)
-    else:
-        cross_check_fennec_apks(apks_metadata_per_paths)
+    package_names_check.validate(apks_metadata_per_paths)
+
+    if not skip_checks_fennec:
+        singular_apk_metadata = list(apks_metadata_per_paths.values())[0]
+        _check_version_matches_package_name(
+            singular_apk_metadata['firefox_version'], singular_apk_metadata['package_name']
+        )
+        _check_apk_package_name(apks_metadata_per_paths, [singular_apk_metadata['package_name']])
+
+        _check_all_apks_have_the_same_firefox_version(apks_metadata_per_paths)
+        _check_all_apks_have_the_same_build_id(apks_metadata_per_paths)
+        _check_all_architectures_and_api_levels_are_present(apks_metadata_per_paths)
+
+    if not skip_check_multiple_locales:
+        _check_all_apks_are_multi_locales(apks_metadata_per_paths)
+
+    if not skip_check_same_locales:
+        _check_all_apks_have_the_same_locales(apks_metadata_per_paths)
+
+    if not skip_check_ordered_version_codes:
+        _check_apks_version_codes_are_correctly_ordered(apks_metadata_per_paths)
 
     logger.info('APKs are sane!')
 
 
-def cross_check_fennec_apks(apks_metadata_per_paths):
-    _check_all_apks_have_the_same_package_name(apks_metadata_per_paths)
-    _check_all_apks_have_the_same_version(apks_metadata_per_paths)
+def _check_apk_package_name(apks_metadata_per_paths, product_types):
+    types = set([metadata['package_name'] for metadata in apks_metadata_per_paths.values()])
 
-    singular_apk_metadata = list(apks_metadata_per_paths.values())[0]
-    _check_version_matches_package_name(
-        singular_apk_metadata['firefox_version'], singular_apk_metadata['package_name']
-    )
-
-    _check_all_apks_have_the_same_build_id(apks_metadata_per_paths)
-    _check_apks_version_codes_are_correctly_ordered(apks_metadata_per_paths)
-
-    _check_all_apks_are_multi_locales(apks_metadata_per_paths)
-    _check_all_apks_have_the_same_locales(apks_metadata_per_paths)
-
-    _check_all_architectures_and_api_levels_are_present(apks_metadata_per_paths)
-
-
-def cross_check_fenix_apks(apks_metadata_per_paths):
-    _check_all_apks_have_the_same_package_name(apks_metadata_per_paths)
-    _check_apks_version_codes_are_correctly_ordered(apks_metadata_per_paths)
-
-
-def cross_check_reference_browser_apks(apks_metadata_per_paths):
-    _check_all_apks_have_the_same_package_name(apks_metadata_per_paths)
-    _check_apks_version_codes_are_correctly_ordered(apks_metadata_per_paths)
-
-
-def cross_check_focus_apks(apks_metadata_per_paths):
-    _check_number_of_distinct_packages(apks_metadata_per_paths, 2)
-    _check_correct_apk_product_types(apks_metadata_per_paths, [PRODUCT.FOCUS, PRODUCT.KLAR])
-
-
-def _check_number_of_distinct_packages(apks_metadata_per_paths, max_packages):
-    all_items = [metadata['package_name'] for metadata in apks_metadata_per_paths.values()]
-    unique_packages = filter_out_identical_values(all_items)
-    if (len(unique_packages) > max_packages):
-        raise BadSetOfApks('Expected max {} package names, found {}'.format(max_packages, len(unique_packages)))
-    logger.info('Found expected number of package names, not more than {}'.format(max_packages))
-
-
-def _check_correct_apk_product_types(apks_metadata_per_paths, product_types):
-    types = set([PRODUCT.get_value_or_none(metadata['package_name']) for metadata in apks_metadata_per_paths.values()])
-    if not types.issubset(product_types):
+    if not types == set(product_types):
         raise BadSetOfApks('Expected product types {}, found {}'.format(product_types, types))
     logger.info('Expected product types {} found'.format(product_types))
 
@@ -91,8 +80,7 @@ def _check_piece_of_metadata_is_unique(key, pretty_key, apks_metadata_per_paths)
     logger.info('All APKs have the same {}: {}'.format(pretty_key, unique_items[0]))
 
 
-_check_all_apks_have_the_same_package_name = partial(_check_piece_of_metadata_is_unique, 'package_name', 'package name')
-_check_all_apks_have_the_same_version = partial(_check_piece_of_metadata_is_unique, 'firefox_version', 'Firefox version')
+_check_all_apks_have_the_same_firefox_version = partial(_check_piece_of_metadata_is_unique, 'firefox_version', 'Firefox version')
 _check_all_apks_have_the_same_build_id = partial(_check_piece_of_metadata_is_unique, 'firefox_build_id', 'Firefox BuildID')
 _check_all_apks_have_the_same_locales = partial(_check_piece_of_metadata_is_unique, 'locales', 'locales')
 
@@ -103,7 +91,7 @@ def _check_version_matches_package_name(version, package_name):
     if (
         (package_name == 'org.mozilla.firefox' and sanitized_version.is_release) or
         # Due to project Dawn, Nightly is now using the Aurora package name. See bug 1357351.
-        (is_package_name_nightly(package_name) and sanitized_version.is_nightly) or
+        (package_name == 'org.mozilla.fennec_aurora' and sanitized_version.is_nightly) or
         (
             # XXX Betas aren't following the regular XX.0bY format. Instead they follow XX.0
             # (which looks like release). Therefore, we can't use sanitized_version.is_beta
@@ -169,9 +157,8 @@ def _check_all_apks_are_multi_locales(apks_metadata_per_paths):
 def _check_all_architectures_and_api_levels_are_present(apks_metadata_per_paths):
     single_metadata = list(apks_metadata_per_paths.values())[0]
     firefox_version = single_metadata['firefox_version']
-    package_name = single_metadata['package_name']
 
-    expected_combos = get_expected_combos(firefox_version, package_name)
+    expected_combos = get_expected_combos(firefox_version)
 
     current_combos = set([
         (metadata['architecture'], metadata['api_level'])