+++ pushapkscript/src/mozapkpublisher/common/apk/extractor.py
@@ -10,7 +10,7 @@ from zipfile import ZipFile
 
 
 from mozapkpublisher.common.exceptions import BadApk, NoLocaleFound
-from mozapkpublisher.common.utils import filter_out_identical_values, PRODUCT
+from mozapkpublisher.common.utils import filter_out_identical_values
 
 from configparser import ConfigParser
 
@@ -26,7 +26,7 @@ _OMNI_JA_LOCATION = 'assets/omni.ja'
 _CHROME_MANIFEST_LOCATION = 'chrome/chrome.manifest'
 
 
-def extract_metadata(original_apk_path):
+def extract_metadata(original_apk_path, extract_firefox_metadata):
     logger.info('Extracting metadata from a copy of "{}"...'.format(original_apk_path))
     metadata = {}
 
@@ -41,14 +41,11 @@ def extract_metadata(original_apk_path):
         metadata['api_level'] = int(androguard_apk.get_min_sdk_version())
         metadata['version_code'] = androguard_apk.get_androidversion_code()
 
-        if PRODUCT.is_focus_flavor(package_name):
-            return metadata
-
         with ZipFile(apk_copy.name) as apk_zip:
             metadata['architecture'] = _extract_architecture(apk_zip, original_apk_path)
             metadata['locales'] = _extract_locales(apk_zip)
 
-            if not (PRODUCT.is_reference_browser(package_name) or PRODUCT.is_fenix(package_name)):
+            if extract_firefox_metadata:
                 metadata['firefox_version'] = _extract_firefox_version(apk_zip)
                 metadata['firefox_build_id'] = _extract_firefox_build_id(apk_zip)
 
@@ -57,8 +54,8 @@ def extract_metadata(original_apk_path):
 
 def _extract_architecture(apk_zip, original_apk_path):
     files_with_architecture_in_path = [
-        file_info.filename for file_info in apk_zip.infolist()
-        if _DIRECTORY_WITH_ARCHITECTURE_METADATA in file_info.filename
+        name for name in apk_zip.namelist()
+        if _DIRECTORY_WITH_ARCHITECTURE_METADATA in name
     ]
 
     if not files_with_architecture_in_path: